--- a/package.json
+++ b/package.json
@@ -33,6 +33,7 @@
     "@types/dateformat": "^1.0.1",
     "@types/mocha": "^5.2.5",
     "@types/node": "^10.7.0",
+    "benchmark": "^2.1.4",
     "chai": "^4.2.0",
     "codecov": "^3.0.4",
     "mocha": "^5.2.0",