--- a/package.json
+++ b/package.json
@@ -26,7 +26,7 @@
   "homepage": "https://github.com/interledgerjs/btp-packet#readme",
   "dependencies": {
     "dateformat": "^3.0.3",
-    "oer-utils": "^3.2.0"
+    "oer-utils": "^4.0.0"
   },
   "devDependencies": {
     "@types/chai": "^4.1.4",